pyproject.toml
[project]
name = "study-buddy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "fastapi[standard]>=0.119.0",
    "pypdf>=5.1.0",
    "python-docx>=1.1.0",
    "openai>=1.0.0",
    "cerebras-cloud-sdk>=1.50.1",
    "python-dotenv>=1.0.0",
    "psycopg[binary]>=3.2.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "psycopg2>=2.9.11",
]
